Cleaned and refactored code - Extracted grouping function and using in all slug material pages - Instead of copying the same function for each material slug page
